+++ Specs/DynamicScene/DynamicBillboardVisualizerSpec.js
@@ -238,7 +238,7 @@ defineSuite([
             if (bb.getShow()) {
                 //Clearing won't actually remove the billboard because of the
                 //internal cache used by the visualizer, instead it just hides it.
-                dynamicObjectCollection.clear();
+                dynamicObjectCollection.removeAll();
                 expect(bb.getShow()).toEqual(false);
                 return true;
             }